리뷰 기능 권한 설정 변경

+++ src/main/java/com/be/friendy/warendy/domain/review/service/ReviewService.java
@@ -5,6 +5,7 @@ import com.be.friendy.warendy.domain.member.repository.MemberRepository;
 import com.be.friendy.warendy.domain.review.dto.request.ReviewUpdateRequest;
 import com.be.friendy.warendy.domain.review.dto.request.WineReviewCreateRequest;
 import com.be.friendy.warendy.domain.review.dto.response.MyReviewSearchResponse;
+import com.be.friendy.warendy.domain.review.dto.response.ReviewUpdateResponse;
 import com.be.friendy.warendy.domain.review.dto.response.WineReviewCreateResponse;
 import com.be.friendy.warendy.domain.review.dto.response.WineReviewSearchByWineIdResponse;
 import com.be.friendy.warendy.domain.review.entity.Review;
@@ -24,37 +25,65 @@ public class ReviewService {
     private final MemberRepository memberRepository;
     private final WineRepository wineRepository;
 
-    public WineReviewCreateResponse createWineReview(
-            Long wineId, WineReviewCreateRequest wineReviewCreateRequest
+    public WineReviewCreateResponse createWineReview(String email
+            , Long wineId, WineReviewCreateRequest wineReviewCreateRequest
     ) {
 
-        Member member = memberRepository
+        Member memberById = memberRepository
                 .findByNickname(wineReviewCreateRequest.getNickname())
                 .orElseThrow(() -> new RuntimeException("user does not exists"));
+        Member memberByEmail = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("user does not exist"));
+
+        if (!memberById.equals(memberByEmail)) {
+            throw new RuntimeException("check the user information");
+        }
 
         Wine wine = wineRepository.findById(wineId)
                 .orElseThrow(() -> new RuntimeException("wine does not exists"));
 
-        return WineReviewCreateResponse.fromEntity(reviewRepository.save(Review.builder()
-                .member(member)
+        return WineReviewCreateResponse.fromEntity(reviewRepository.save(Review
+                .builder()
+                .member(memberById)
                 .wine(wine)
                 .contents(wineReviewCreateRequest.getContents())
                 .rating(wineReviewCreateRequest.getRating())
                 .build()));
     }
 
-    public Review updateReview(
-            Long reviewId, ReviewUpdateRequest reviewUpdateRequest
+    public ReviewUpdateResponse updateReview(
+            String email, Long reviewId, ReviewUpdateRequest reviewUpdateRequest
     ) {
+        Member memberById = memberRepository
+                .findByNickname(reviewUpdateRequest.getNickname())
+                .orElseThrow(() -> new RuntimeException("user does not exists"));
+        Member memberByEmail = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("user does not exist"));
+        if (!memberById.equals(memberByEmail)) {
+            throw new RuntimeException("check the user information");
+        }
+
         Review nowReview = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new RuntimeException("the review does not exists"));
         nowReview.updateReviewInfo(reviewUpdateRequest);
-        return reviewRepository.save(nowReview);
+        reviewRepository.save(nowReview);
+        return ReviewUpdateResponse.fromEntity(nowReview);
     }
 
-    public void deleteReview(Long reviewId) {
+    public void deleteReview(String email, Long reviewId) {
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new RuntimeException("the review does not exists"));
+        Member memberById = memberRepository
+                .findByNickname(review.getMember().getNickname())
+                .orElseThrow(() -> new RuntimeException("user does not exists"));
+
+        Member memberByEmail = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("user does not exist"));
+
+        if (!memberById.equals(memberByEmail)) {
+            throw new RuntimeException("check the user information");
+        }
+
         reviewRepository.save(review);      // - test 코드 검증 위해
         reviewRepository.delete(review);
     }
@@ -69,11 +98,17 @@ public class ReviewService {
     }
 
     public Page<MyReviewSearchResponse> searchMyReview(
-            String nickname, Pageable pageable
+            String email, String nickname, Pageable pageable
     ) {
-        Member member = memberRepository.findByNickname(nickname)
+        Member memberById = memberRepository.findByNickname(nickname)
                 .orElseThrow(() -> new RuntimeException("user does not exist"));
-        return reviewRepository.findByMember(member, pageable)
+        Member memberByEmail = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("user does not exist"));
+
+        if (!memberById.equals(memberByEmail)) {
+            throw new RuntimeException("check the user information");
+        }
+        return reviewRepository.findByMember(memberById, pageable)
                 .map(MyReviewSearchResponse::fromEntity);
     }
